+++ visual-js/visual-nightwatch/src/nightwatch/commands/sauceVisualCheck.ts
@@ -1,10 +1,18 @@
-const EventEmitter = require('events').EventEmitter;
-const { ensureError, getFullPageConfig } = require('@saucelabs/visual');
-const { parseIgnoreOptions, toIgnoreRegionIn } = require('../../utils/regions');
-const { getMetaInfo, getVisualApi } = require('../../utils/api');
-const { VISUAL_BUILD_ID_KEY } = require('../../utils/constants');
+import { ensureError, getFullPageConfig, isSkipMode } from '@saucelabs/visual';
+import { parseIgnoreOptions, toIgnoreRegionIn } from '../../utils/regions';
+import { getMetaInfo, getVisualApi } from '../../utils/api';
+import { VISUAL_BUILD_ID_KEY } from '../../utils/constants';
+import { NightwatchAPI } from 'nightwatch';
+import { NightwatchCustomCommandsModel } from 'nightwatch/types/custom-command';
+import { CheckOptions, RunnerSettings } from '../../types';
+import { Runnable } from 'mocha';
 
-module.exports = class SauceVisualCheck extends EventEmitter {
+type APIType = NightwatchAPI & {
+  capabilities: Record<string, any>;
+  options: RunnerSettings;
+};
+
+export default class SauceVisualCheck implements NightwatchCustomCommandsModel {
   // These are used for Cucumber
   static featureName = '';
   static scenarioName = '';
@@ -12,15 +20,28 @@ module.exports = class SauceVisualCheck extends EventEmitter {
   // Mocha context can be passed, but
   //  - not for all runners,
   //  - not always as the third argument
-  async command(name, optionsArg = {}, mochaContextArg = {}) {
+  async command(name: string, optionsArg = {}, mochaContextArg = {}) {
     console.log(`Checking ${name}`);
-    const nightwatchBrowserObject = this.api;
+    if (isSkipMode()) {
+      console.log(`Checking ${name}: SKIPPED`);
+      global.skipped = (global.skipped ?? 0) + 1;
+      return null;
+    }
+
+    // @ts-expect-error API doesn't allow us to type the options / extra values from webdriver.
+    const nightwatchBrowserObject: APIType = this.api as unknown as APIType;
     //
     // Check if the first argument is a Mocha context
-    const isMochaContext = (arg) => arg.title && arg.ctx?._runnable?.title;
-    const [options, mochaContext] = isMochaContext(optionsArg)
-      ? [{}, optionsArg]
-      : [optionsArg, mochaContextArg];
+    const isMochaContext = (arg: Runnable | Record<any, any>) =>
+      arg.title &&
+      // TODO: Replace with a non-private member / usage.
+      // @ts-ignore
+      arg.ctx?._runnable?.title;
+    const [options, mochaContext] = (
+      isMochaContext(optionsArg)
+        ? [{}, optionsArg]
+        : [optionsArg, mochaContextArg]
+    ) as [CheckOptions, Runnable];
     //
     // Getting the suite and testname from the current test
     let module = '';
@@ -36,6 +57,8 @@ module.exports = class SauceVisualCheck extends EventEmitter {
     const suiteName =
       mochaContext?.title || SauceVisualCheck.featureName || module || '';
     const testName =
+      // TODO: Replace with a non-private member / usage.
+      // @ts-ignore
       mochaContext?.ctx?._runnable?.title ||
       SauceVisualCheck.scenarioName ||
       defaultTestName ||
@@ -54,7 +77,7 @@ module.exports = class SauceVisualCheck extends EventEmitter {
       capabilities,
       sessionId,
       options: {
-        webdriver: { host, port },
+        webdriver: { host, port } = {},
         sauceVisualService: {
           captureDom: globalCaptureDom = false,
           fullPage,
@@ -72,14 +95,15 @@ module.exports = class SauceVisualCheck extends EventEmitter {
 
     if (!buildId) {
       nightwatchBrowserObject.assert.fail('No buildId found');
-      this.emit('complete', null);
-      return;
+      return null;
     }
 
     const api = getVisualApi(sauceConfig);
     const metaInfo = await getMetaInfo(api, sessionId, jobId);
 
-    let result = null;
+    let result: Awaited<
+      ReturnType<typeof api.createSnapshotFromWebDriver>
+    > | null = null;
 
     try {
       result = await api.createSnapshotFromWebDriver({
@@ -104,6 +128,6 @@ module.exports = class SauceVisualCheck extends EventEmitter {
       nightwatchBrowserObject.assert.fail(errorMessage);
     }
 
-    this.emit('complete', result);
+    return result;
   }
-};
+}